refactor logic to support a computeHeightForDefaultText() method
